Api request now will use User API Key instead of Development Key Added PostClaimRequest API

+++ app/src/main/java/com/truesightid/data/source/local/entity/ClaimEntity.kt
@@ -37,5 +37,5 @@ data class ClaimEntity(
     var downvote: Int,
 
     @ColumnInfo(name = "date")
-    var date: Int
+    var date: Float
 ) : Parcelable
